test: regroup unit tests
